test change requests for py2 and py3 unicode versions
